New: `WriteBus.contracomposeWriter`
Allows people to provide an operator to transform the write-bus stream.

`contracomposeWriter` is more general than `contramapWriter`, as the first takes
an operator on streams and the second a mapping function its events.

The advantage of having `contracomposeWriter` is that it provides a way for
people to combine streams.

Now `contramapWriter` is defined in terms of `contracomposeWriter`.