New: Recursive, depth-first transaction queue
Fixes #39 by changing transaction scheduling logic and delaying the evaluation of arguments to Var.update. Essentially, with these changes, when you schedule two transactions in a row, you know that the first transaction will complete, and the second transaction will not start executing until all of the transactions scheduled while running the first transaction are executed. And this logic is recursive. I will document the details later, if this logic proves itself.