API: Var.update and Var.updater now send error into unhandled instead of throwing inside a transaction
Throwing inside a transaction has unpredictable results, because you do not know when the transaction will actually execute. If you try to update a failed var outside a transaction, e.g. in a repl, the transaction will run immediately and will throw. But if you call var.update while a transaction is running (e.g. in an Observer), the transaction will only be executed after the current transaction has finished, so your var.update call will not throw.