Fix: Bug in JsPriorityQueue implementation that produced glitches in complex cases.
Specifically, when two or more observables were pending in the same transaction at the same time, if one of them actually synchronously depended on the other, it could fire ahead of the observable that it depended on, causing a glitch. Whether this actually happened also depended on the order of (internal) observers in the observable graph, you could avoid this bug just by luck. This bug could affect nested combineWith or delaySync observables, resulting in extraneous events (glitches) in case of combineWith, or observables firing in the wrong order (in case of delaySync).

See https://gitter.im/Laminar_/Lobby?at=6007655b36db01248a8bf5a9 for the original bug report.