API: Reimplement Future-based observables in terms of js.Promise
This removes the dependency on Future (unless you actually want to use it) and standardizes behaviour to the lowest common denominator which is js.Promise. In previous code, calling fromFuture(promise.future) and fromJsPromise(promise) would produce different timings because it is impossible to observe a JS Promise's value synchronously, and promise.future implicitly inherits that property.