Fix: MergeStream was not firing events in a new transaction
Correct behaviour is to fire the first event in the same transaction, and any subsequent events received in the same transaction in a new transaction. It was introduced in 28b5c04785473afcb833679d7d832f4e39377027 and documentation was written for it, but then in 28b5c04785473afcb833679d7d832f4e39377027 we started firing all events in the new transaction, and in 28dbf23979b4444662983dcb3fdd15bc89b87f15 we started firing all events in the original transaction. The last change definitely appears unintentional as it is rolled up into an unrelated refactor.

All the original tests pass, and I have now restored the behaviour that was promised in the docs. I added a new test to verify that using mergeWith with combineWith does not cause glitches. As expected, both incorrect implementtations fail this test in different ways, whereas the correct one passes.